+++ core-ng/src/test/java/core/framework/util/FilesTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 
 import java.nio.file.Path;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author neo
@@ -12,7 +12,15 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 class FilesTest {
     @Test
     void tempFile() {
-        Path path = Files.tempFile();
-        assertTrue(path.toString().endsWith(".tmp"));
+        Path file = Files.tempFile();
+        assertThat(file.getFileName().toString()).endsWith(".tmp");
+    }
+
+    @Test
+    void tempDir() {
+        Path tempDir = Files.tempDir();
+        assertThat(tempDir).exists().isDirectory();
+
+        Files.deleteDir(tempDir);
     }
 }